chor: refactored code
